객체미용체조 7원칙 - 2개 이상의 원시타입 프로퍼티를 갖는 타입 금지
