+++ ParcelInvoiceMaker/ParcelInvoiceMaker/ParcelProcessor.swift
@@ -7,33 +7,13 @@
 import Foundation
 
 class ParcelInformation {
-    let address: String
-    var receiverName: String
-    var receiverMobile: String
-    let deliveryCost: Int
-    private let discount: Discount
-    var discountedCost: Int {
-        switch discount {
-        case .none:
-            return deliveryCost
-        case .vip:
-            return deliveryCost / 5 * 4
-        case .coupon:
-            return deliveryCost / 2
-        }
-    }
+    let reciverInformation: ReceiverInformation
+    let costInformation: CostInformation
 
     init(address: String, receiverName: String, receiverMobile: String, deliveryCost: Int, discount: Discount) {
-        self.address = address
-        self.receiverName = receiverName
-        self.receiverMobile = receiverMobile
-        self.deliveryCost = deliveryCost
-        self.discount = discount
+        self.reciverInformation = ReceiverInformation(address: address, name: receiverName, mobile: receiverMobile)
+        self.costInformation = CostInformation(deliveryCost: deliveryCost, discount: discount)
     }
-}
-
-enum Discount: Int {
-    case none = 0, vip, coupon
 }
 
 class ParcelOrderProcessor {
